reorganize and document inductor extension points
